Move browser/hash out of index files
