+++ packages/browser/types/browser.d.ts
@@ -1,0 +1,4 @@
+import { ResponseHandler } from "@hickory/root";
+import { BrowserHistoryOptions, BrowserHistory } from "./types";
+export * from "./types";
+export declare function browser(fn: ResponseHandler, options?: BrowserHistoryOptions): BrowserHistory;
