+++ packages/hash/src/hash.ts
