--- packages/hash/types/index.d.ts
@@ -1,4 +1,0 @@
-import { ResponseHandler } from "@hickory/root";
-import { HashOptions, HashHistory } from "./types";
-export * from "./types";
-export declare function hash(fn: ResponseHandler, options?: HashOptions): HashHistory;
